+++ example/lib/app/theme.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+// ignore: avoid_classes_with_only_static_members
+class AppTheme {
+  static final light = ThemeData(
+    brightness: Brightness.light,
+    colorScheme: const ColorScheme.light(secondary: Colors.red),
+    disabledColor: Colors.grey.shade400,
+    visualDensity: VisualDensity.adaptivePlatformDensity,
+  );
+
+  static final dark = ThemeData(
+    brightness: Brightness.dark,
+    colorScheme: const ColorScheme.dark(secondary: Colors.red),
+    disabledColor: Colors.grey.shade400,
+    visualDensity: VisualDensity.adaptivePlatformDensity,
+  );
+}