+++ lib/src/helpers/adaptive_controls.dart
@@ -1,0 +1,31 @@
+import 'package:chewie/chewie.dart';
+import 'package:chewie/src/material/material_desktop_controls.dart';
+import 'package:flutter/material.dart';
+
+class AdaptiveControls extends StatelessWidget {
+  const AdaptiveControls({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    switch (Theme.of(context).platform) {
+      case TargetPlatform.android:
+      case TargetPlatform.fuchsia:
+        return const MaterialControls();
+
+      case TargetPlatform.macOS:
+      case TargetPlatform.windows:
+      case TargetPlatform.linux:
+        return const MaterialDesktopControls();
+
+      case TargetPlatform.iOS:
+        return const CupertinoControls(
+          backgroundColor: Color.fromRGBO(41, 41, 41, 0.7),
+          iconColor: Color.fromARGB(255, 200, 200, 200),
+        );
+      default:
+        return const MaterialControls();
+    }
+  }
+}